+++ src/components/subscribe-to-unlock-card.tsx
@@ -21,7 +21,6 @@ interface Props {
 export const SubscribeToUnlockCard = ({
   onSubscribe,
   loadingSubscribe,
-  subscribeDisabled,
   post,
 }: Props) => {
   // const [loadingTrial, setLoadingTrial] = useState(false);
@@ -80,7 +79,7 @@ export const SubscribeToUnlockCard = ({
           sx={{ width: '100%', py: 1.5 }}
           onClick={onSubscribe}
           loading={loadingSubscribe}
-          disabled={subscribeDisabled}
+          // disabled={subscribeDisabled}
         >
           <IconPlayerPlay size={20} style={{ marginRight: 5 }} />
           Join